+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Require golang 1.17 for `go generate`, removing the binary dependencies from go.mod.
 - Update CSI Spec to 1.5.0
 
+### Fixed
+
+- ControllerUnpublishVolume could fail if LINSTOR node was deleted before the detach operation was executed.
+
 ## [0.17.0] - 2021-12-09
 
 ### Added