+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A new parameter `usePvcName` was added. When "true", the driver will (try to) use the PVC name + namespace instead
   of the generated PV name as ID. Because this might generate name conflicts in certain scenarios, it remains disabled
   by default.
+- Include [`linstor-wait-until`] in the image, to make it easy to wait for the LINSTOR API before starting the driver.
+
+[`linstor-wait-until`]: https://github.com/LINBIT/linstor-wait-until
 
 ### Changed