+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.18.0] - 2022-02-24
 
 ### Added
 
@@ -360,7 +360,8 @@ on storage with the following (case sensitive!) options:<!-- Needs Docs -->
 - all storageClass parameters options are now case insenstive
 - non-debug logging is less verbose in general
 
-[Unreleased]: https://github.com/piraeusdatastore/linstor-csi/compare/v0.17.0...HEAD
+[Unreleased]: https://github.com/piraeusdatastore/linstor-csi/compare/v0.18.0...HEAD
+[0.18.0]: https://github.com/piraeusdatastore/linstor-csi/compare/v0.17.0...v0.18.0
 [0.17.0]: https://github.com/piraeusdatastore/linstor-csi/compare/v0.16.1...v0.17.0
 [0.16.1]: https://github.com/piraeusdatastore/linstor-csi/compare/v0.16.0...v0.16.1
 [0.16.0]: https://github.com/piraeusdatastore/linstor-csi/compare/v0.15.1...v0.16.0